requirements.txt
# Appium and Selenium dependencies
Appium-Python-Client==3.1.0
selenium==4.15.2

# Testing framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0

# Reporting
allure-pytest==2.13.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1

# Logging and utilities
colorlog==6.8.0

